Added OrderController, OrderService, OrderRepository. Created OrderDto and OrderStatusDto. Also added mapper for OrderDto. + some refactoring.

+++ ISHop/src/main/java/by/it/academy/ishop/dtos/ProductDto.java
@@ -22,6 +22,6 @@ public class ProductDto {
     @Positive
     private BigDecimal price;
     @Positive
-    private int number;
+    private Long number;
     private String description;
 }
